+++ src/Symfony/Component/DomCrawler/CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
 * Added `Crawler::matches()` method.
 * Added `Crawler::closest()` method.
 * Added `Crawler::outerHtml()` method.
+* Added an argument to the `Crawler::text()` method to opt-in normalizing whitespaces.
 
 4.3.0
 -----